Pomazanov_Aleksandr/feedback-analysis-agent/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
pandas>=2.0.0
pydantic>=2.0.0
click>=8.0.0
python-dotenv>=1.0.0
rich>=13.0.0